add --force option
