+++ src/cli.ts
@@ -2,20 +2,20 @@ import * as fs from "fs";
 import * as path from "path";
 import * as commander from "commander";
 import { ConsoleLogger } from "@akashic/akashic-cli-commons";
-import { promiseExportZip } from "./bundle";
+import { promiseExportZip } from "./exportZip";
 
-interface CommandParameterObject {
+export interface CommandParameterObject {
 	cwd?: string;
 	quiet?: boolean;
 	output?: string;
+	force?: boolean;
 	strip?: boolean;
 	minify?: boolean;
 	bundle?: boolean;
 	hashFilename?: number | boolean;
 }
 
-
-function cli(param: CommandParameterObject): void {
+export function cli(param: CommandParameterObject): void {
 	var logger = new ConsoleLogger({ quiet: param.quiet });
 	Promise.resolve()
 		.then(() => promiseExportZip({
@@ -24,6 +24,7 @@ function cli(param: CommandParameterObject): void {
 			strip: param.strip,
 			source: param.cwd,
 			dest: param.output,
+			force: param.force,
 			hashLength: !param.hashFilename ? 0 : (param.hashFilename === true) ? 20 : Number(param.hashFilename),
 			logger
 		}))
@@ -43,6 +44,7 @@ commander
 	.option("-C, --cwd <dir>", "A directory containing a game.json (default: .)")
 	.option("-q, --quiet", "Suppress output")
 	.option("-o, --output <fileName>", "Name of output file (default: game.zip)")
+	.option("-f, --force", "Overwrites existing files")
 	.option("-s, --strip", "Contain only files refered by game.json")
 	.option("-M, --minify", "Minify JavaScript files")
 	.option("-H, --hash-filename [length]", "Rename asset files with their hash values")
@@ -54,6 +56,7 @@ export function run(argv: string[]): void {
 		cwd: commander["cwd"],
 		quiet: commander["quiet"],
 		output: commander["output"],
+		force: commander["force"],
 		strip: commander["strip"],
 		minify: commander["minify"],
 		hashFilename: commander["hashFilename"],
